+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added NextAuth to project
 - Added session handling
 - Added [custom next Document](pages/_document.tsx)
+- Added Inter font
+- Added React Bootstrap Icons to dependencies
 
 ### Changed